refaactor rename datamodel to datasource in src ref #29
