+++ src/database-server/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/data-model)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/data-source)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/data-access)
 
 find_package(Boost 1.67 REQUIRED system thread)
