+++ src/database-server/client/client.h
@@ -46,7 +46,7 @@
 
 #include <messages.pb.h>
 
-#include "src/database-server/data-model/idatamodel.h"
+#include "src/database-server/data-source/idatamodel.h"
 
 class Client : public QObject {
   Q_OBJECT
